refactor(package): move packages
- 도메인 내에 너무 광범위한 관심사의 클래스들이 모여 가독성이 떨어지는 현상 개선
