+++ src/main/java/nextstep/subway/line/domain/stationAdapter/SafeStation.java
@@ -1,4 +1,4 @@
-package nextstep.subway.line.domain;
+package nextstep.subway.line.domain.stationAdapter;
 
 public interface SafeStation {
     SafeStationInfo getStationSafely(Long stationId);
