+++ src/main/java/nextstep/subway/line/domain/stationAdapter/SafeStationAdapter.java
@@ -1,4 +1,4 @@
-package nextstep.subway.line.domain;
+package nextstep.subway.line.domain.stationAdapter;
 
 import nextstep.subway.line.domain.exceptions.StationNotFoundException;
 import nextstep.subway.station.application.StationService;
