+++ src/test/java/nextstep/subway/line/application/LineServiceTest.java
@@ -3,6 +3,8 @@ package nextstep.subway.line.application;
 import nextstep.subway.line.application.exceptions.LineNotFoundException;
 import nextstep.subway.line.domain.*;
 import nextstep.subway.line.domain.exceptions.StationNotFoundException;
+import nextstep.subway.line.domain.stationAdapter.SafeStationAdapter;
+import nextstep.subway.line.domain.stationAdapter.SafeStationInfo;
 import nextstep.subway.line.dto.LineRequest;
 import nextstep.subway.line.dto.LineResponse;
 import nextstep.subway.line.dto.StationInLineResponse;
